refactor: move utils into /lib
