+++ src/components/modals/ClassInfo/ClassInfo.tsx
@@ -14,9 +14,9 @@ import { useUserConfig } from "@/lib/hooks/useUserConfig";
 import { useUserSessions } from "@/lib/hooks/useUserSessions";
 import { getCapitalizedWeekday, isClassInThePast } from "@/lib/integration/common";
 import { stringifyClassPopularity } from "@/lib/popularity";
+import { formatNameArray } from "@/lib/utils/arrayUtils";
+import { hexWithOpacityToRgb } from "@/lib/utils/colorUtils";
 import { ClassPopularity, RezervoClass, SessionStatus, StatusColors, UserNameWithIsSelf } from "@/types/rezervo";
-import { formatNameArray } from "@/utils/arrayUtils";
-import { hexWithOpacityToRgb } from "@/utils/colorUtils";
 
 export default function ClassInfo({
     _class,
